refactor: 사각형과 원을 Figure로 추상화
